+++ commands/random.js
@@ -1,8 +1,10 @@
 // Get a random anime
+const kitsu = require("node-kitsu");
+const VNDB = require("vndb-api");
+const moment = require("moment");
 
 exports.run = async (client, message, args, level) => {
-  const kitsu = require("node-kitsu");
-  if (!args[0]) return message.channel.send("Missing argument. Please specify either anime or manga.");
+  if (!args[0]) return message.channel.send("Missing argument. Please specify \"anime\", \"manga\", or \"VN\".");
   var embed;
   var msg;
   var found = false;
@@ -14,14 +16,14 @@ exports.run = async (client, message, args, level) => {
         try {
           var results = await kitsu.listAnime(rnd);
         } catch (ex) {
-          message.channel.send("An error occured running this command. This is likely due to an issue on Kitsu's end, and not an error with the bot. Please try your command again later.");
+          message.channel.send("An error occurred running this command. Please try again later.");
           found = true;
           return client.logger.error(`An error occurred with the command: ${ex}`);
         }
         try {
           var aniresult = results[0].attributes;
           if (!aniresult.titles) throw "No result found";
-          embed = { "embed": {
+          embed = {
             "title": aniresult.canonicalTitle || aniresult.titles.en || aniresult.titles.en_jp,
             "url": `https://kitsu.io/anime/${aniresult.slug}`,
             "description": client.cleanSyn(aniresult.synopsis),
@@ -32,12 +34,12 @@ exports.run = async (client, message, args, level) => {
               { "name": "Episodes:", "value":  `${aniresult.episodeCount || 0} (${aniresult.subtype})`, "inline": true },
               { "name": "Status:", "value": aniresult.status == "tba" ? "TBA" : `${aniresult.status.charAt(0).toUpperCase()}${aniresult.status.substr(1).toLowerCase()}`, "inline": true }
             ]
-          } };
+          };
           client.logger.log(`Anime sucesfully generated: ${embed.embed.title}`);
           found = true;
         } catch (ex) {
           // ¯\_(ツ)_/¯
-          client.logger.warn("Randomizing failed. Retrying.");
+          client.logger.warn("Randomizing anime failed. Retrying.");
           await client.wait(1000);
         }
       }
@@ -49,14 +51,14 @@ exports.run = async (client, message, args, level) => {
         try {
           var results = await kitsu.listManga(rnd);
         } catch (ex) {
-          message.channel.send("An error occured running this command. This is likely due to an issue on Kitsu's end, and not an error with the bot. Please try your command again later.");
+          message.channel.send("An error occurred running this command. Please try again later.");
           found = true;
           return client.logger.error(`An error occurred with the command: ${ex}`);
         }
         try {
           var aniresult = results[0].attributes;
           if (!aniresult.titles) throw "No result found";
-          embed = { "embed": {
+          embed = {
             "title": aniresult.canonicalTitle || aniresult.titles.en || aniresult.titles.en_jp,
             "url": `https://kitsu.io/manga/${aniresult.slug}`,
             "description": client.cleanSyn(aniresult.synopsis),
@@ -67,20 +69,55 @@ exports.run = async (client, message, args, level) => {
               { "name": "Chapters:", "value":  `${aniresult.episodeCount || 0} (${aniresult.subtype})`, "inline": true },
               { "name": "Status:", "value": aniresult.status == "tba" ? "TBA" : `${aniresult.status.charAt(0).toUpperCase()}${aniresult.status.substr(1).toLowerCase()}`, "inline": true }
             ]
-          } };
+          };
           client.logger.log(`Manga sucesfully generated: ${embed.embed.title}`);
           found = true;
         } catch (ex) {
           // ¯\_(ツ)_/¯
-          client.logger.warn("Randomizing failed. Retrying.");
+          client.logger.warn("Randomizing manga failed. Retrying.");
+          await client.wait(1000);
+        }
+      }
+      break;
+    case "vn":
+      var msg = await message.channel.send("Please wait a moment...");
+      while (!found) {
+        try {
+          const vndb = new VNDB("michelle-vndb");
+          const dbinfo = await vndb.query("dbstats");
+          var rnd = client.randInt(1,dbinfo.vn);
+          var results = await vndb.query(`get vn basic,details (id = ${rnd})`);
+        } catch (ex) {
+          message.channel.send("An error occurred running this command. Please try again later.");
+          return client.logger.error(`An error occurred with the command:\n${JSON.stringify(ex)}`);
+        }
+        try {
+          var vnresult = results.items[0];
+          if (!vnresult.title) throw "No result found";
+          embed = {
+            "title": vnresult.title,
+            "url": `https://vndb.org/v${vnresult.id}`,
+            "description": client.cleanSyn(vnresult.description),
+            "color": client.colorInt("#071c30"),
+            "image": {"url": vnresult.image_nsfw ? (message.channel.nsfw ? vnresult.image : "https://michelle.jacenboy.com/assets/nsfw-overlay.png") : vnresult.image},
+            "fields": [
+              {"name": "Release Date", "value": moment(vnresult.released).format("MMM D[,] YYYY"), "inline": true},
+              {"name": "Languages", "value": vnresult.languages.join(", "), "inline": true},
+              {"name": "Platforms", "value": vnresult.platforms.join(", "), "inline": true}
+            ]
+          };
+          found = true;
+        } catch (ex) {
+          // ¯\_(ツ)_/¯
+          client.logger.warn("Randomizing VN failed. Retrying.");
           await client.wait(1000);
         }
       }
       break;
     default:
-      return message.channel.send("Invalid argument. Please specify either anime or manga.");
+      return message.channel.send("Invalid argument. Please specify \"anime\", \"manga\", or \"VN\".");
   }
-  msg.edit(embed);
+  msg.edit({"content": "", "embed": embed});
 };
 
 exports.conf = {